refactor(cohortCreation): move getAllCriteriaItems to utils to avoid circular imports and update tests accordingly
